Применил SwiftLint
