+++ ClearCodeGB/ClearCodeGB/App/Flows/ViewController.swift
@@ -11,9 +11,6 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
+        view.backgroundColor = .red
     }
-
-
 }
-
