+++ ClearCodeGB/ClearCodeGB/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
