+++ ClearCodeGB/ClearCodeGB/Resources/Assets.xcassets/Contents.json
