+++ ClearCodeGB/Podfile
@@ -1,0 +1,21 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '9.0'
+
+target 'ClearCodeGB' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for ClearCodeGB
+
+pod 'SwiftLint'
+
+  target 'ClearCodeGBTests' do
+    inherit! :search_paths
+    # Pods for testing
+  end
+
+  target 'ClearCodeGBUITests' do
+    # Pods for testing
+  end
+
+end
